backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
motor==3.6.0
pymongo==4.9.2
python-multipart==0.0.9
openai-whisper==20240930
numpy==1.26.4
plotly==5.24.1
python-dotenv==1.0.1
pytest==8.3.3
